requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cramjam==2.8.1
Django==5.0.2
fastparquet==2024.2.0
fsspec==2024.2.0
idna==3.6
joblib==1.3.2
Naked==0.1.32
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pyarrow==15.0.0
pycryptodome==3.20.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
threadpoolctl==3.3.0
tqdm==4.66.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.1
